Limit transitive search to included dependencies
The AbstractResolveDependencies class, when building a list of
transitive Maven dependencies to check, will only include trees in the
DependencyNode#INCLUDED state. This will ensure only the coalesced
version of an artifact is used, when there is a conflict between
dependencies with the same groupId and artifactId, but different
versions.

This fixes the BanDuplicateClasses to no longer detect conflicts between
multiple versions of the same GA combination.