Also support 1.x version schemes
Despite https://openjdk.java.net/jeps/223 officially deprecating 1.x
schemes, we should be permissive here to avoid crashes.

For instance, in Jenkins the version in use is `1.${java.level}`.
https://github.com/jenkinsci/plugin-pom/blob/master/pom.xml#L522

So supporting both formats will avoid unnecessary headaches around.
We'll see possibly in the future to stop supporting this when the oldest
line is >=9 and no more 1.x format can exist.